+++ core/src/jsMain/kotlin/JsNavigationNode.kt
@@ -5,7 +5,10 @@ import kotlinx.browser.document
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
+import org.w3c.dom.Element
+import org.w3c.dom.HTMLElement
 import org.w3c.dom.MutationObserver
+import org.w3c.dom.MutationObserverInit
 
 abstract class JsNavigationNode<Config : Base, Base : NavigationNodeDefaultConfig>(
     override val chain: NavigationChain<Base>,
@@ -15,6 +18,9 @@ abstract class JsNavigationNode<Config : Base, Base : NavigationNodeDefaultConfi
     override val configState: StateFlow<Config> = _configState.asStateFlow()
     protected val htmlElementOrThrow
         get() = configState.value.htmlElementOrThrow
+    private val _htmlElementStateFlow = MutableStateFlow<HTMLElement?>(null)
+    protected val htmlElementStateFlow = _htmlElementStateFlow.asStateFlow()
+
     override var config: Config
         get() = _configState.value
         set(value) { _configState.value = value }
@@ -22,30 +28,25 @@ abstract class JsNavigationNode<Config : Base, Base : NavigationNodeDefaultConfi
     protected var observer: MutationObserver? = null
 
     override fun onResume() {
-        runCatching {
-            htmlElementOrThrow
-        }.onSuccess {
-            super.onResume()
-        }.onFailure {
-            observer = MutationObserver { _, mutationObserver ->
-                runCatching {
-                    htmlElementOrThrow
-                }.onSuccess {
-                    if (state == NavigationNodeState.RESUMED) {
-                        super.onResume()
-                    }
-                    mutationObserver.disconnect()
-                }
-            }.apply {
-                observe(document.body ?: return)
-            }
+        super.onResume()
+        inline fun refresh() {
+            _htmlElementStateFlow.value = runCatching {
+                htmlElementOrThrow
+            }.getOrNull()
         }
+        observer = MutationObserver { _, _ ->
+            refresh()
+        }.apply {
+            observe(document, MutationObserverInit(childList = true, subtree = true, attributes = true))
+        }
+        refresh()
     }
 
     override fun onPause() {
         super.onPause()
         observer ?.disconnect()
         observer = null
+        _htmlElementStateFlow.value = null
     }
 
 }