+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 0.0.11
 
+* `JS`:
+  * `JsNavigationNode` got `htmlElementStateFlow` which is automatically updated when node is in resume state
+
 ## 0.0.10
 
 * `Versions`: